Add a physical file lock for truststore write access

+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/utils/CertificateMgtUtils.java
@@ -114,38 +114,43 @@ public class CertificateMgtUtils {
                 log.error("Certificate is empty for the provided alias " + alias);
                 return ResponseCode.INTERNAL_SERVER_ERROR;
             }
+            String tempTrustStore = TRUST_STORE + ".temp.lock";
+            if (TrustStoreUtils.acquireLockWithRetries(tempTrustStore)) {
+                //Read the client-truststore.jks into a KeyStore.
+                File trustStoreFile = new File(TRUST_STORE);
+                localTrustStoreStream = new FileInputStream(trustStoreFile);
+                KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
+                TrustStoreUtils.loadCerts(trustStore, TRUST_STORE,TRUST_STORE_PASSWORD);
 
-            //Read the client-truststore.jks into a KeyStore.
-            File trustStoreFile = new File(TRUST_STORE);
-            localTrustStoreStream = new FileInputStream(trustStoreFile);
-            KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
-            TrustStoreUtils.loadCerts(trustStore, TRUST_STORE,TRUST_STORE_PASSWORD);
-
-            CertificateFactory cf = CertificateFactory.getInstance(CERTIFICATE_TYPE);
-            while (serverCert.available() > 0) {
-                Certificate certificate = cf.generateCertificate(serverCert);
-                //Check whether the Alias exists in the trust store.
-                if (trustStore.containsAlias(alias)) {
-                    isCertExists = true;
-                } else {
-                    /*
-                    * If alias is not exists, check whether the certificate is expired or not. If expired set the
-                    * expired flag.
-                    * */
-                    X509Certificate x509Certificate = (X509Certificate) certificate;
-                    if (x509Certificate.getNotAfter().getTime() <= System.currentTimeMillis()) {
-                        expired = true;
-                        if (log.isDebugEnabled()) {
-                            log.debug("Provided certificate is expired.");
-                        }
+                CertificateFactory cf = CertificateFactory.getInstance(CERTIFICATE_TYPE);
+                while (serverCert.available() > 0) {
+                    Certificate certificate = cf.generateCertificate(serverCert);
+                    //Check whether the Alias exists in the trust store.
+                    if (trustStore.containsAlias(alias)) {
+                        isCertExists = true;
                     } else {
-                        //If not expired add the certificate to trust store.
-                        trustStore.setCertificateEntry(alias, certificate);
+                        /*
+                         * If alias is not exists, check whether the certificate is expired or not. If expired set the
+                         * expired flag.
+                         * */
+                        X509Certificate x509Certificate = (X509Certificate) certificate;
+                        if (x509Certificate.getNotAfter().getTime() <= System.currentTimeMillis()) {
+                            expired = true;
+                            if (log.isDebugEnabled()) {
+                                log.debug("Provided certificate is expired.");
+                            }
+                        } else {
+                            //If not expired add the certificate to trust store.
+                            trustStore.setCertificateEntry(alias, certificate);
+                        }
                     }
                 }
+                fileOutputStream = new FileOutputStream(trustStoreFile);
+                trustStore.store(fileOutputStream, TRUST_STORE_PASSWORD);
+            } else {
+                log.error("Could not acquire lock to update the trust store.");
+                return ResponseCode.INTERNAL_SERVER_ERROR;
             }
-            fileOutputStream = new FileOutputStream(trustStoreFile);
-            trustStore.store(fileOutputStream, TRUST_STORE_PASSWORD);
             responseCode = expired ? ResponseCode.CERTIFICATE_EXPIRED :
                     isCertExists ? ResponseCode.ALIAS_EXISTS_IN_TRUST_STORE : ResponseCode.SUCCESS;
         } catch (CertificateException e) {
@@ -165,6 +170,9 @@ public class CertificateMgtUtils {
             responseCode = ResponseCode.INTERNAL_SERVER_ERROR;
         } catch (IOException e) {
             log.error("Error in loading the certificate.", e);
+            responseCode = ResponseCode.INTERNAL_SERVER_ERROR;
+        } catch (InterruptedException e) {
+            log.error("Error in acquiring lock to update the trust store.", e);
             responseCode = ResponseCode.INTERNAL_SERVER_ERROR;
         } finally {
             closeStreams(localTrustStoreStream, fileOutputStream, serverCert);
@@ -245,23 +253,29 @@ public class CertificateMgtUtils {
 
         boolean isExists; //Check for the existence of the certificate in trust store.
         try {
-            File trustStoreFile = new File(TRUST_STORE);
-            localTrustStoreStream = new FileInputStream(trustStoreFile);
-            KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
-            TrustStoreUtils.loadCerts(trustStore, TRUST_STORE,TRUST_STORE_PASSWORD);
+            String tempTrustStore = TRUST_STORE + ".temp.lock";
+            if (TrustStoreUtils.acquireLockWithRetries(tempTrustStore)) {
+                File trustStoreFile = new File(TRUST_STORE);
+                localTrustStoreStream = new FileInputStream(trustStoreFile);
+                KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
+                TrustStoreUtils.loadCerts(trustStore, TRUST_STORE,TRUST_STORE_PASSWORD);
 
-            if (trustStore.containsAlias(alias)) {
-                trustStore.deleteEntry(alias);
-                isExists = true;
-            } else {
-                isExists = false;
-                if (log.isDebugEnabled()) {
-                    log.debug("Certificate for alias '" + alias + "' not found in the trust store.");
+                if (trustStore.containsAlias(alias)) {
+                    trustStore.deleteEntry(alias);
+                    isExists = true;
+                } else {
+                    isExists = false;
+                    if (log.isDebugEnabled()) {
+                        log.debug("Certificate for alias '" + alias + "' not found in the trust store.");
+                    }
                 }
-            }
 
-            fileOutputStream = new FileOutputStream(trustStoreFile);
-            trustStore.store(fileOutputStream, TRUST_STORE_PASSWORD);
+                fileOutputStream = new FileOutputStream(trustStoreFile);
+                trustStore.store(fileOutputStream, TRUST_STORE_PASSWORD);
+            } else {
+                log.error("Could not acquire lock to update the trust store.");
+                return ResponseCode.INTERNAL_SERVER_ERROR;
+            }
             responseCode = isExists ? ResponseCode.SUCCESS : ResponseCode.CERTIFICATE_NOT_FOUND;
         } catch (IOException e) {
             log.error("Error in loading the certificate.", e);
@@ -274,6 +288,9 @@ public class CertificateMgtUtils {
             responseCode = ResponseCode.INTERNAL_SERVER_ERROR;
         } catch (KeyStoreException e) {
             log.error("Error reading certificate contents.", e);
+            responseCode = ResponseCode.INTERNAL_SERVER_ERROR;
+        } catch (InterruptedException e) {
+            log.error("Error in acquiring lock to update the trust store.", e);
             responseCode = ResponseCode.INTERNAL_SERVER_ERROR;
         } finally {
             closeStreams(localTrustStoreStream, fileOutputStream);
